pyproject.toml
[project]
name = "english-tg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "groq>=0.31.0",
    "openai>=1.101.0",
    "python-dotenv>=1.1.1",
]
